refactor: CostCategory 모임 관련 adapter 패키지 이동
